feat: fetch content detail by route id from API (#1471)
